requirements.txt
beautifulsoup4==4.11.1
bleach==4.1.0
bokeh==2.4.2
datashape==0.5.4
matplotlib==3.5.1
matplotlib-inline==0.1.6
numpy==1.21.5
numpydoc==1.2
pandas==1.4.4
pandas-profiling==3.3.1
Pillow==9.0.1
regex==2022.3.15
scikit-learn==1.0.2
scipy==1.7.3
Scrapy==2.6.1
seaborn==0.11.2